resume-helper/backend/requirements.txt
# 后端依赖
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0

# 数据库
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20.0

# LLM
litellm>=1.40.0

# 文档解析
pdfplumber>=0.11.0
python-docx>=1.1.0
# pytesseract>=0.3.10  # OCR（可选，需安装Tesseract）

# 模板与导出
jinja2>=3.1.0
docxtpl>=0.18.0
playwright>=1.40.0

# Agent编排
langgraph>=0.2.0

# 向量检索（Phase 6）
# chromadb>=0.5.0

# 工具
python-multipart>=0.0.9
aiofiles>=24.0
